Add safety checks.

+++ mfw-setup.sh
@@ -18,25 +18,48 @@ if [ ! -f $SETTINGS_BACKUP_FILE ]; then
   cp $SETTINGS_FILE $SETTINGS_BACKUP_FILE
 fi
 
+UPDATED=0
 read -p "LAN IP Address: " LAN_IP
+if [ "$LAN_IP" != "" ] ; then
+    sed -i -r \
+        's/("v4StaticAddress":) "[^"]+"/\1 "'$LAN_IP'"/' \
+        $SETTINGS_FILE
+    UPDATED=1
+fi
+
 read -p "LAN Network Prefix: " LAN_PREFIX
+if [ "$LAN_PREFIX" != "" ] ; then
+    sed -i -r \
+        's/("v4StaticPrefix":) \d+/\1 '$LAN_PREFIX'/' \
+        $SETTINGS_FILE
+    UPDATED=1
+fi
 
-sed -i -r \
- 's/("v4StaticAddress":) "[^"]+"/\1 "'$LAN_IP'"/; s/("v4StaticPrefix":) \d+/\1 '$LAN_PREFIX'/' \
- $SETTINGS_FILE
+read -p "Enable SSH on WAN? (y/N): " SSH_WAN
+if [ "$SSH_WAN" == "y" ] ; then
+    sed -i -r \
+        '$!N;s/"enabled": false,(\n\s+"description": "Accept SSH on WANs)/"enabled": true,\1/;P;D' \
+        $SETTINGS_FILE
+    UPDATED=1
+fi
 
-sed -i -r \
- '$!N;s/"enabled": false,(\n\s+"description": "Accept SSH on WANs)/"enabled": true,\1/;P;D' \
- $SETTINGS_FILE
-
-sync-settings
+if [ $UPDATED -eq 1 ] ; then
+    sync-settings
+fi
 
 ##
 ## Update storage
 ##
-opkg update
-opkg install fdisk block-mount rsync
+mount | grep overlay
+OVERLAY_EXISTS=$?
 
+if [ $OVERLAY_EXISTS -eq 1 ]; then 
+	opkg update
+	opkg install fdisk block-mount rsync
+
+	##
+	## Create new partion on second disk.
+	##
 echo "
 n
 p
@@ -47,38 +70,40 @@ i 1
 w
 " | fdisk /dev/sdb
 
-mkfs.ext4 /dev/sdb1
+	mkfs.ext4 /dev/sdb1
 
-##
-## Setup new overlay.
-##
-uuid=$(block info | \
-	grep sdb1 | \
-	sed -r 's/[[:alnum:]]+=/\n&/g' | \
-	awk -F= '$1=="UUID"{print $2}' | \
-	sed -e 's/"//g' \
-)
-echo "uuid=$uuid"
+	##
+	## Setup new overlay.
+	##
+	uuid=$(block info | \
+		grep sdb1 | \
+		sed -r 's/[[:alnum:]]+=/\n&/g' | \
+		awk -F= '$1=="UUID"{print $2}' | \
+		sed -e 's/"//g' \
+	)
+	echo "uuid=$uuid"
 
-uci -q delete fstab.overlay
-uci set fstab.overlay="mount"
-uci set fstab.overlay.uuid=$uuid
-uci set fstab.overlay.target="/overlay"
-uci commit fstab
+	uci -q delete fstab.overlay
+	uci set fstab.overlay="mount"
+	uci set fstab.overlay.uuid=$uuid
+	uci set fstab.overlay.target="/overlay"
+	uci commit fstab
 
-mount /dev/sdb1 /mnt
-cp -f -a /overlay/. /mnt
-umount /mnt
+	mount /dev/sdb1 /mnt
+	cp -f -a /overlay/. /mnt
+	umount /mnt
 
-##
-## opkg to overlay
-##
-OPKG_CONF_FILE=/etc/opkg.conf
-sed -i -e \
- '/^lists_dir\s/s:/var/opkg-lists$:/usr/lib/opkg/lists:' \
- $OPKG_CONF_FILE
-opkg update
+	##
+	## opkg to overlay
+	##
+	OPKG_CONF_FILE=/etc/opkg.conf
+	sed -i -e \
+ 		'/^lists_dir\s/s:/var/opkg-lists$:/usr/lib/opkg/lists:' \
+ 		$OPKG_CONF_FILE
+	opkg update
 
-reboot
+fi
+
+#reboot
